cli: Refactor mount module into separate files
Split src/mount.rs into a module directory structure:
- src/mount/mod.rs: shared types (MountOpts, MountHandle) and public API
- src/mount/fuse.rs: FUSE backend implementation
- src/mount/nfs.rs: NFS backend implementation
